requirements.txt
# Core
python-dotenv==0.21.1
loguru==0.7.2
pydantic==2.5.2

# Telegram Bot
aiogram==3.3.0

# RAG Pipeline
langchain==0.1.5
langchain-community==0.0.17
langchain-openai==0.0.5
openai==1.10.0

# Vector Store
faiss-cpu==1.7.4

# Document Loaders
pypdf==4.0.1
python-docx==1.1.0

# Google Drive
PyDrive2==1.17.0
google-api-python-client==2.115.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
